feat(store): deepen Config parsing — Zod StoreConfig with fs+env adapters
- Extract StoreConfig to src/store-config.ts with Zod schema (single source for 5 fields)
- fs adapter: yaml simple key:value + stripQuotes/inlineComment, per-key Zod validation + warn
- env adapter: DSH_BETTER_EDIT_STORE_DIR/AUTO_GITIGNORE true|false case-insensitive, normalized + warn
- Merge defaults + yaml + env, final Zod parse + expand + fallback central, mtime+env cache
- paths.ts thin adapter re-exports StoreConfig/loadConfig/_resetConfigCache

Refs: candidate 5 speculative
